Disable automatic `drop` of `OutStream`
`OutStream` contains a `*mut` pointer to the stream struct which
causes it to be dropped as soon as another `OutStream` struct is created
from the same pointer (one mutable or n readable pointer). My `drop`
implementation called `soundio_out_stream_destroy` which must now be
done manually by calling the `destroy` method when everything is
finished. Streams are created on devices, so the original device could
manage a references to the streams and destroy them automatically as
soon as the device is dropped.